+++ entities/teacher.rb
@@ -1,0 +1,14 @@
+require_relative './person'
+
+class Teacher < Person
+  attr_accessor :specialization
+
+  def initialize(age, specialization, name = 'unknown', parent_permission: true)
+    super(id, name, age, parent_permission)
+    @specialization = specialization
+  end
+
+  def can_use_services?
+    true
+  end
+end
